Use commit authors as committers, too
If `user.name` and `user.email` are not configured in the setup kirby-autogit is used in,
`$this->commit(...)` will fail, because git prints its infamous "tell me who you are"
message.

By explicitly providing the two config keys with each commit command,
we can use the determined user as the committer and thus have the commit
work anyway without needing any changes to any of git's configs.